+++ app/src/main/java/com/jesusr00/primarosdas/models/Event.kt
@@ -1,4 +1,4 @@
 package com.jesusr00.primarosdas.models
 
 
-class Event (val day: Int, val startTime: Int, val endTime: Int, val title: String, val description: String?, val location: String?, val type: String?)
+class Event (val day: Int, val startTime: Long, val endTime: Long, val title: String, val description: String?, val location: String?, val type: String?)
